fix problem cause the programm will not run
fix the problem of would not running  because the program cannot find "repo" directory. which is mentioned in the arduboy forum but no one fixed it.
After fix the repo should be at /home/pi/arduboy/repo
and no limit to where uploader.exe is but it's recommended to be at /home/pi/arduboy/uploader.exe to avoid mess up files.